+++ packages/web-app-files/src/mixins/routes.js
@@ -5,6 +5,7 @@ import {
   isLightweightRoute,
   isPublicFilesRoute,
   isPublicPage,
+  isProjectsRoute,
   isSharedWithMeRoute,
   isSharedWithOthersRoute,
   isTrashbinRoute
@@ -35,6 +36,9 @@ export default {
     },
     isPublicFilesRoute() {
       return isPublicFilesRoute(this.$route)
+    },
+    isProjectsRoute() {
+      return isProjectsRoute(this.$route)
     },
     isPublicPage() {
       return isPublicPage(this.$route)