+++ packages/web-app-files/src/views/SharedWithMe.vue
@@ -47,7 +47,7 @@
             </div>
           </template>
           <template #contextMenu="{ resource }">
-            <context-actions :item="resource" />
+            <context-actions v-if="isHighlightedFile(resource)" :item="resource" />
           </template>
           <!--<template v-if="pendingHasMore" #footer>
             <div class="uk-width-1-1 uk-text-center oc-mt">
@@ -135,7 +135,7 @@
           </div>
         </template>
         <template #contextMenu="{ resource }">
-          <context-actions :item="resource" />
+          <context-actions v-if="isHighlightedFile(resource)" :item="resource" />
         </template>
         <template #footer>
           <list-info
@@ -421,6 +421,10 @@ export default {
 
     togglePendingShowMore() {
       this.showMorePending = !this.showMorePending
+    },
+
+    isHighlightedFile(resource) {
+      return resource && resource.id === this.highlightedFile?.id
     }
   }
 }