+++ packages/web-app-files/tests/integration/specs/peopleExpirationDate.spec.js
@@ -12,6 +12,8 @@ import sdkMock from '@/__mocks__/sdk'
 
 import FileShares from '@files/src/components/SideBar/Shares/FileShares.vue'
 
+const formatDate = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' }
+
 const existingShares = [
   {
     shareInfo: {
@@ -60,14 +62,14 @@ describe('Users can set expiration date when sharing with users or groups', () =
 
     const newDate = getDateInFuture(2)
     const dateSelector = document.evaluate(
-      `//span[contains(@class, "vc-day-content vc-focusable") and text()="${newDate.getDate()}"]`,
+      `//span[contains(@class, "vc-day-content vc-focusable") and not(contains(@class, "is-disabled")) and text()="${newDate.getDate()}"]`,
       baseElement,
       null,
       XPathResult.FIRST_ORDERED_NODE_TYPE,
       null
     ).singleNodeValue
 
-    if (dateSelector.classList.contains('is-disabled')) {
+    if (!dateSelector) {
       const nextMonthBtn = baseElement.querySelector('.vc-arrow.is-right')
 
       userEvent.click(nextMonthBtn)
@@ -125,27 +127,42 @@ describe('Users can set expiration date when sharing with users or groups', () =
 
     expect(editDialog).toBeVisible()
     expect(getByTestId('recipient-datepicker')).toBeVisible()
+    expect(within(getByTestId('recipient-datepicker')).getByText('Expires in 2 days')).toBeVisible()
+
     await fireEvent.click(getByTestId('recipient-datepicker-btn'))
 
     const newDate = getDateInFuture(4)
-    const dateSelector = document.evaluate(
-      `//span[contains(@class, "vc-day-content vc-focusable") and text()="${newDate.getDate()}"]`,
+    let dateSelector = document.evaluate(
+      `//span[contains(@class, "vc-day-content vc-focusable") and @tabindex="-1" and @aria-label="${newDate.toLocaleDateString(
+        'en',
+        formatDate
+      )}" and text()="${newDate.getDate()}"]`,
       baseElement,
       null,
       XPathResult.FIRST_ORDERED_NODE_TYPE,
       null
     ).singleNodeValue
 
-    if (dateSelector.classList.contains('is-disabled')) {
+    if (!dateSelector) {
       const nextMonthBtn = baseElement.querySelector('.vc-arrow.is-right')
 
-      userEvent.click(nextMonthBtn)
+      await userEvent.click(nextMonthBtn)
+      dateSelector = document.evaluate(
+        `//span[contains(@class, "vc-day-content vc-focusable") and @tabindex="-1" and @aria-label="${newDate.toLocaleDateString(
+          'en',
+          formatDate
+        )}" and text()="${newDate.getDate()}"]`,
+        baseElement,
+        null,
+        XPathResult.FIRST_ORDERED_NODE_TYPE,
+        null
+      ).singleNodeValue
     }
 
     expect(
       await findByText(newDate.toLocaleString('en', { month: 'long', year: 'numeric' }))
     ).toBeVisible()
-    userEvent.click(dateSelector)
+    await userEvent.click(dateSelector)
     expect(await within(editDialog).findByText('Expires in 4 days')).toBeVisible()
     await fireEvent.click(getByTestId('recipient-edit-btn-save'))
     await waitFor(() => expect(queryByTestId('recipient-dialog-edit')).toBe(null))
@@ -276,30 +293,37 @@ describe('Users can set expiration date when sharing with users or groups', () =
 
     const userInAutocomplete = await findByTestId('recipient-autocomplete-item-bob')
     expect(userInAutocomplete).toBeVisible()
-    userEvent.click(userInAutocomplete)
+    await userEvent.click(userInAutocomplete)
     expect(await findByTestId('recipient-container-bob')).toBeVisible()
     expect(getByTestId('recipient-datepicker')).toBeVisible()
     await fireEvent.click(getByTestId('recipient-datepicker-btn'))
 
     const newDate = getDateInFuture(2)
-    const dateSelector = document.evaluate(
-      `//span[contains(@class, "vc-day-content vc-focusable") and text()="${newDate.getDate()}"]`,
+    let dateSelector = document.evaluate(
+      `//span[contains(@class, "vc-day-content vc-focusable") and @tabindex="-1" and not(contains(@class, "is-disabled")) and text()="${newDate.getDate()}"]`,
       baseElement,
       null,
       XPathResult.FIRST_ORDERED_NODE_TYPE,
       null
     ).singleNodeValue
 
-    if (dateSelector.classList.contains('is-disabled')) {
-      const nextMonthBtn = baseElement.querySelector('.vc-arrow.is-right')
+    if (!dateSelector) {
+      const nextMonthBtn = baseElement.querySelector('.vc-arrow.is-left') // the date is less than default expiration date so needs to go back
+      await userEvent.click(nextMonthBtn)
 
-      userEvent.click(nextMonthBtn)
+      dateSelector = document.evaluate(
+        `//span[contains(@class, "vc-day-content vc-focusable") and not(contains(@class, "is-disabled")) and text()="${newDate.getDate()}"]`,
+        baseElement,
+        null,
+        XPathResult.FIRST_ORDERED_NODE_TYPE,
+        null
+      ).singleNodeValue
     }
 
     expect(
       await findByText(newDate.toLocaleString('en', { month: 'long', year: 'numeric' }))
     ).toBeVisible()
-    userEvent.click(dateSelector)
+    await userEvent.click(dateSelector)
     expect(await findByText('Expires in 2 days')).toBeVisible()
 
     const shareBtn = getByTestId('new-collaborator-share-btn')
@@ -369,24 +393,37 @@ describe('Users can set expiration date when sharing with users or groups', () =
     await fireEvent.click(getByTestId('recipient-datepicker-btn'))
 
     const newDate = getDateInFuture(4)
-    const dateSelector = document.evaluate(
-      `//span[contains(@class, "vc-day-content vc-focusable") and text()="${newDate.getDate()}"]`,
+    let dateSelector = document.evaluate(
+      `//span[contains(@class, "vc-day-content vc-focusable") and @tabindex="-1" and @aria-label="${newDate.toLocaleDateString(
+        'en',
+        formatDate
+      )}" and text()="${newDate.getDate()}"]`,
       baseElement,
       null,
       XPathResult.FIRST_ORDERED_NODE_TYPE,
       null
     ).singleNodeValue
 
-    if (dateSelector.classList.contains('is-disabled')) {
+    if (!dateSelector) {
       const nextMonthBtn = baseElement.querySelector('.vc-arrow.is-right')
 
-      userEvent.click(nextMonthBtn)
+      await userEvent.click(nextMonthBtn)
+      dateSelector = document.evaluate(
+        `//span[contains(@class, "vc-day-content vc-focusable") and @tabindex="-1" and @aria-label="${newDate.toLocaleDateString(
+          'en',
+          formatDate
+        )}" and text()="${newDate.getDate()}"]`,
+        baseElement,
+        null,
+        XPathResult.FIRST_ORDERED_NODE_TYPE,
+        null
+      ).singleNodeValue
     }
 
     expect(
       await findByText(newDate.toLocaleString('en', { month: 'long', year: 'numeric' }))
     ).toBeVisible()
-    userEvent.click(dateSelector)
+    await userEvent.click(dateSelector)
     expect(await within(editDialog).findByText('Expires in 4 days')).toBeVisible()
     await fireEvent.click(getByTestId('recipient-edit-btn-save'))
     await waitFor(() => expect(queryByTestId('recipient-dialog-edit')).toBe(null))