+++ packages/web-runtime/src/store/apps.js
@@ -46,13 +46,13 @@ const mutations = {
     const editor = {
       app,
       icon: extension.icon,
+      img: extension.img,
       newTab: extension.newTab || false,
       routeName: extension.routeName,
       routes: extension.routes || [],
       extension: extension.extension,
       handler: extension.handler,
-      showInRightClickMenu: extension.showInRightClickMenu || false,
-      canBeDefault: 'canBeDefault' in extension ? extension.canBeDefault : true
+      canBeDefault: 'canBeDefault' in extension? extension.canBeDefault : true
     }
 
     state.fileEditors.push(editor)
@@ -97,7 +97,7 @@ const mutations = {
       name: appInfo.name || appInfo.id,
       id: appInfo.id,
       icon: appInfo.icon || 'check_box_outline_blank',
-      iconImg: appInfo.iconImg
+      img: appInfo.img || null
     }
     state.meta[app.id] = app
   },